+++ index.js
@@ -9,6 +9,9 @@
 
 var Q = require('q')
 
+// Store for promises created during a template execution
+var promises = null
+
 /**
  * Returns a new Handlebars instance that
  * * allows helpers to return promises
@@ -22,6 +25,7 @@ var Q = require('q')
  * @param {string=} options.placeholder
  * @returns {Handlebars} a modified Handlebars object
  */
+
 module.exports = function promisedHandlebars (Handlebars, options) {
   options = options || {}
   options.placeholder = options.placeholder || '\u0001'
@@ -29,16 +33,23 @@ module.exports = function promisedHandlebars (Handlebars, options) {
   // one line from from substack's quotemeta-package
   var regex = new RegExp(String(options.placeholder).replace(/(\W)/g, '\\$1'), 'g')
 
-  var promises = null
   var engine = Handlebars.create()
 
   // Wrap `registerHelper` with a custom function
   var oldRegisterHelper = engine.registerHelper
-  engine.registerHelper = function (keyOrObject, fn) {
-
+  engine.registerPromiseHelper = function (keyOrObject, fn) {
     if (typeof keyOrObject === 'string') {
       // Register a custom helper-function instead of actual helper
       oldRegisterHelper.call(this, keyOrObject, function () {
+        // Wrap "optons.fn" and "options.inverse" to apply the same
+        var helperOpts = arguments[arguments.length - 1]
+        if (helperOpts.fn) {
+          helperOpts.fn = wrapAndResolve(helperOpts.fn)
+        }
+        if (helperOpts.inverse) {
+          helperOpts.inverse = wrapAndResolve(helperOpts.inverse)
+        }
+
         var result = Q(fn.apply(this, arguments))
         // Remember promise and return placeholder instead
         promises.push(result)
@@ -49,7 +60,7 @@ module.exports = function promisedHandlebars (Handlebars, options) {
       // Call `registerHelper` again for each object
       // This simulates the default Handlebars-behaviour
       Object.keys(keyOrObject).forEach(function (key) {
-        engine.registerHelper(key, keyOrObject[key])
+        engine.registerPromiseHelper(key, keyOrObject[key])
       })
     }
   }
@@ -57,13 +68,17 @@ module.exports = function promisedHandlebars (Handlebars, options) {
   var oldCompile = engine.compile
   engine.compile = function () {
     var fn = oldCompile.apply(this, arguments)
-    // Wrap the compiled function
+    return wrapAndResolve(fn)
+  // Wrap the compiled function
+  }
+
+  function wrapAndResolve (fn) {
     return function () {
       if (promises) {
         // "promises" array already exists: We are executing a partial
         // That means we are called from somewhere within Handlebars.
         // Handlebars does not like promises, so we act as normal as possible.
-        return fn.apply(engine, arguments)
+        return fn.apply(undefined, arguments)
       }
       try {
         // We are called from outside Handlebars (initial call)
@@ -73,10 +88,10 @@ module.exports = function promisedHandlebars (Handlebars, options) {
         promises = []
         // Execute template (helpers are getting called and store promises
         // into the array
-        var resultWithPlaceholders = fn.apply(engine, arguments)
+        var resultWithPlaceholders = fn.apply(undefined, arguments)
         return Q.all(promises).then(function (results) {
           // Promises are fulfilled. Insert real values into the result.
-          return resultWithPlaceholders.replace(regex, function () {
+          return String(resultWithPlaceholders).replace(regex, function () {
             return results.shift()
           })
         })
@@ -88,5 +103,6 @@ module.exports = function promisedHandlebars (Handlebars, options) {
     }
 
   }
+
   return engine
 }